Phase 13: Move Branding into Admin sub-tab
- Remove Branding nav item from settings/layout.tsx sidebar
- Extract BrandingTab into shared component (components/settings/branding-tab.tsx)
- BrandingPage imports BrandingTab from shared component
- Add 'Branding' as tab in admin/page.tsx (Users | Instance settings | Branding)
