+++ apps/web/app/(dashboard)/settings/admin/page.tsx
@@ -14,6 +14,7 @@ import { useAuthStore } from "@/stores/auth-store";
 import { useRouter } from "next/navigation";
 import type { User, UserStatus } from "@/types";
 import { InstanceSettingsTab } from "@/components/settings/instance-settings-tab";
+import { BrandingTab } from "@/components/settings/branding-tab";
 
 function BulkInviteDialog() {
   const [open, setOpen] = React.useState(false);
@@ -167,7 +168,7 @@ function userStatusBadge(status: UserStatus) {
 export default function AdminPage() {
   const { user, isSuperAdmin } = useAuthStore();
   const router = useRouter();
-  const [tab, setTab] = React.useState<"users" | "instance">("users");
+  const [tab, setTab] = React.useState<"users" | "instance" | "branding">("users");
 
   const { data: usersResp, isLoading: loadingUsers } = useSWR<User[]>(
     isSuperAdmin ? "/admin/users" : null,
@@ -249,7 +250,7 @@ export default function AdminPage() {
 
       {/* Sub-tabs */}
       <div className="flex gap-1 border-b border-border">
-        {([["users", "Users"], ["instance", "Instance settings"]] as const).map(([key, label]) => (
+        {([["users", "Users"], ["instance", "Instance settings"], ["branding", "Branding"]] as const).map(([key, label]) => (
           <button
             key={key}
             onClick={() => setTab(key)}
@@ -266,6 +267,7 @@ export default function AdminPage() {
       </div>
 
       {tab === "instance" && <InstanceSettingsTab />}
+      {tab === "branding" && <BrandingTab />}
 
       {/* User management */}
       {tab === "users" && (
